src/requirements.txt
bokeh==1.1.0
Click==7.0
Flask==1.0.2
Flask-MQTT==1.0.5
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
numpy==1.16.2
packaging==19.0
paho-mqtt==1.4.0
Pillow==6.0.0
pyparsing==2.4.0
python-dateutil==2.8.0
PyYAML==5.1
six==1.12.0
tornado==6.0.2
typing==3.6.6
Werkzeug==0.15.2